[DEV-12530] Properly select the right pluralization rule depending on locale
