+++ packages/intl/src/formatting/datetime/lib.ts
@@ -1,21 +1,8 @@
 import { Locale } from '../../locales';
+import { withCache } from '../../utils';
 
 import { toDate } from './shared';
 import type { Iso8601Date, Timezone, UnixTimestampInSeconds } from './types';
-
-function withCache<P extends any[], T extends any>(fn: (...args: P) => T) {
-    const CACHE = new Map<string, T>();
-
-    return (...args: P) => {
-        const key = JSON.stringify(args);
-        const cached = CACHE.get(key);
-        if (cached) return cached;
-
-        const value = fn(...args);
-        CACHE.set(key, value);
-        return value;
-    };
-}
 
 const getDateFormat = withCache(
     (locale: Locale.Code, timeZone: string, options: Intl.DateTimeFormatOptions = {}) =>
