+++ packages/intl/src/utils/index.ts
@@ -1,0 +1,1 @@
+export { withCache } from './withCache';
